Fix supervisord version
In b70a855dbfa05d5db7edc9ffa5a8167d642b0cfe supervisord's version was
set to what was mistakenly assumed to be the latest version available
to Alpine. It wasn't, and it caused errors on Docker image build. This
change corrects the Dockerfile to use the proper latest version.